+++ elements/button.jsx
@@ -6,7 +6,6 @@
  * @param  {String} [props.className] Optional.
  * @param  {Object} [props.divProps] Optional.
  * @param  {String} [props.divProps.$] Any property applicable to a &lt;div&gt; tag. If "className" supplied, override "props.className".
- * @param  {Function} [props.handleClick] Optional.
  * @param  {Node|Node[]} [props.children] Optional.
  * @return {Node}
  */
@@ -14,7 +13,7 @@ ReactSUI.Button = (props) => {
   let className = ReactSUI.utils.addClass("ui button", props.className);
 
   return (
-    <div className={className} {...props.divProps} onClick={props.handleClick}>
+    <div className={className} {...props.divProps}>
       {props.children}
     </div>
   );
@@ -23,6 +22,5 @@ ReactSUI.Button = (props) => {
 ReactSUI.Button.propTypes = {
   className  : React.PropTypes.string,
   divProps   : React.PropTypes.object,
-  handleClick: React.PropTypes.func,
   children   : React.PropTypes.node
 };