+++ elements/image.jsx
@@ -4,7 +4,6 @@
  * @memberOf ReactSUI
  * @param  {Object} props
  * @param  {String} [props.className] Optional.
- * @param  {String} props.src
  * @param  {Object} [props.imgProps] Optional.
  * @param  {String} [props.imgProps.$] Any property applicable to an &lt;img&gt; tag. If "className", "src" supplied, override "props.className", "props.src".
  * @return {Node}
@@ -12,11 +11,10 @@
 ReactSUI.Image = (props) => {
   let className = ReactSUI.utils.addClass("ui image", props.className);
 
-  return <img className={className} src={props.src} {...props.imgProps} />;
+  return <img className={className} {...props.imgProps} />;
 };
 
 ReactSUI.Image.propTypes = {
   className: React.PropTypes.string,
-  src      : React.PropTypes.string.isRequired,
   imgProps : React.PropTypes.object
 };